Break up simulation file into three parts. With test updates.
