+++ src/simulation.js
@@ -1,716 +1,21 @@
-import * as d3 from 'd3';
-
-/**
- * Configuration constants for all graph visualization.
- * Consolidates settings for main histograms, preview histograms, and task row mini graphs.
- */
-const GRAPH_CONFIG = {
-  // Main histogram settings
-  histogram: {
-    barCutoff: 600, // Switch to scatter plot above this threshold
-    maxBuckets: 120, // Maximum buckets for preview histograms
-    width: 800, // Default histogram width
-    height: 500, // Default histogram height
-    margin: { // SVG margins
-      top: 10,
-      right: 30,
-      bottom: 50,
-      left: 60,
-    },
-  },
-  // Task row mini graph settings
-  miniGraph: {
-    width: 140,
-    height: 26,
-    maxBuckets: 24,
-    gap: 1, // Gap between bars
-  },
-};
-
-/**
- * Immutable snapshot of the original default graph configuration values.
- * Use this in reset operations rather than repeating the literal defaults.
- */
-const GRAPH_CONFIG_DEFAULTS = Object.freeze({
-  histogram: Object.freeze({
-    barCutoff: 600,
-    maxBuckets: 120,
-    width: 800,
-    height: 500,
-    margin: Object.freeze({
-      top: 10, right: 30, bottom: 50, left: 60,
-    }),
-  }),
-  miniGraph: Object.freeze({
-    width: 140,
-    height: 26,
-    maxBuckets: 24,
-    gap: 1,
-  }),
-});
-
-/**
- * Get a random number is a given range.
- * @param {int} minimum
- * @param {int} maximum
- * @returns int
- */
-function getRandom(minimum, maximum) {
-  const min = Math.ceil(minimum);
-  const max = Math.floor(maximum);
-  return Math.floor(Math.random() * ((max - min) + 1)) + min;
-}
-
-/**
- * Calculates the upper bound for a specific task record. Because developers
- * are known to underestimate, the idea here is the less confidence they have
- * in their estimate the more risk that it could go much higher than expected.
- * So for every 10% drop in confidence we add the max estimate on again.
- * 90% leaves the upper bound at max estimate.
- * 80% gives us max * 2.
- * 70% max * 3.
- * And so on.
- * @param {*} maxEstimate
- * @param {*} confidence
- * @returns Integer
- */
-function taskUpperBound(maxEstimate, confidence) {
-  // Calculate multiplier based on confidence level
-  // 100% conf = 1x, 90% = 1x, 80% = 2x, 70% = 3x, etc.
-  const confidencePercent = Math.round(confidence * 100);
-  const multiplier = Math.max(1, Math.ceil((100 - confidencePercent) / 10));
-  const boundary = maxEstimate * multiplier;
-  return boundary;
-}
-
-/**
- * Calculates the lower bound for a specific task record. Mirrors taskUpperBound:
- * the less confidence in the estimate, the further an underrun can drop below min.
- * 90% leaves the lower bound at min estimate (no further underrun possible).
- * 80% gives us min / 2.
- * 70% min / 3.
- * And so on.
- * @param {number} minEstimate
- * @param {number} confidence
- * @returns {number}
- */
-function taskLowerBound(minEstimate, confidence) {
-  const confidencePercent = Math.round(confidence * 100);
-  const multiplier = Math.max(1, Math.ceil((100 - confidencePercent) / 10));
-  return minEstimate / multiplier;
-}
-
-/**
- * Does the estimate for one task. It picks a random number between min and
- * max confidence % of the time. If the number is outside the range, overruns
- * receive 75% of the outside-confidence budget and underruns receive 25%,
- * reflecting that tasks more often run over than under. Underruns are bounded
- * by a confidence-scaled floor: the lower the confidence, the further below
- * min the result can reach (mirroring the overrun upper-bound logic). For
- * every 10% drop in confidence the overrun ceiling grows by 100% of max,
- * and the underrun floor drops by 1/multiplier of min.
- * @param {number} minimum
- * @param {number} maximum
- * @param {number} confidence
- * @returns {number}
- */
-function generateEstimate(minimum, maximum, confidence) {
-  const max = parseFloat(maximum);
-  const min = parseFloat(minimum);
-  const base = getRandom(1, 1000);
-  const boundary = confidence * 1000;
-  // Underruns get 25% of the outside-confidence budget; overruns get 75%.
-  const underrunBudget = Math.floor((1000 - boundary) * 0.25);
-  const range = max - min;
-  const maxOverrun = taskUpperBound(max, confidence);
-  const minUnderrun = taskLowerBound(min, confidence);
-  let total = 0;
-
-  if (base < boundary) {
-    // Generate random value within the min-max range
-    total = (Math.random() * range) + min;
-  } else if ((base - boundary) < underrunBudget) {
-    // Underrun: confidence-scaled floor up to min
-    total = min === 0 ? 0 : minUnderrun + (Math.random() * (min - minUnderrun));
-  } else {
-    // Overrun: random value between max and maxOverrun
-    total = max + (Math.random() * (maxOverrun - max));
-  }
-
-  // Round to whole numbers (integers) for cleaner histogram storage
-  return Math.round(total);
-}
-
-/**
- * Counts the total number of values in this result set.
- * @param {Array} data Array of summarized results.
- * @returns the count of values in the result set.
- */
-function getValueCount(data) {
-  // Find total count of values.
-  let valueCount = 0;
-  data.forEach((value) => {
-    valueCount += value;
-  });
-  return valueCount;
-}
-
-/**
- * Gaussian kernel function for KDE.
- * @param {number} distance Distance from the point
- * @param {number} bandwidth Bandwidth parameter
- * @returns {number} Kernel value
- */
-function gaussianKernel(distance, bandwidth) {
-  const u = distance / bandwidth;
-  return Math.exp(-0.5 * u * u) / Math.sqrt(2 * Math.PI);
-}
-
-/**
- * Calculates kernel density estimate for visualization.
- * @param {Array} data Array of frequency values
- * @param {number} minBin Starting index
- * @param {number} maxBin Ending index
- * @returns {Array} Array of KDE values
- */
-function calculateKDE(data, minBin, maxBin) {
-  // Calculate bandwidth using Silverman's rule of thumb
-  const totalCount = getValueCount(data);
-  const range = maxBin - minBin;
-  const bandwidth = Math.max(range * 0.02, 1);
-
-  const kdeValues = [];
-  const samplePoints = Math.min(200, range); // Limit sample points for performance
-  const step = range / samplePoints;
-
-  for (let sample = 0; sample < samplePoints; sample += 1) {
-    const x = minBin + (sample * step);
-    let density = 0;
-
-    // Sum kernel contributions from all data points
-    data.forEach((frequency, i) => {
-      if (frequency > 0) {
-        const dataPoint = i + minBin;
-        const distance = x - dataPoint;
-        // Weight by frequency (how many times this value occurred)
-        density += frequency * gaussianKernel(distance, bandwidth);
-      }
-    });
-
-    // Normalize by total count and bandwidth
-    kdeValues.push(density / (totalCount * bandwidth));
-  }
-
-  // Scale KDE values to match the frequency scale for visualization
-  const maxKDE = Math.max(...kdeValues);
-  const maxFreq = Math.max(...data);
-  const scaleFactor = maxFreq / maxKDE;
-
-  return kdeValues.map((v) => v * scaleFactor);
-}
-
-/**
- * Calculates the median value for all times run during a series of simulations. In the expected
- * array each cell is the number of times the result was equal to the index.
- * @param {Array} data Array of summarized results.
- * @returns the median from value of the run.
- */
-function getMedian(data) {
-  // Find total count of values.
-  const valueCount = getValueCount(data);
-
-  // Handle empty dataset (all zeros in histogram)
-  if (valueCount === 0) {
-    return 0;
-  }
-
-  // Walk back to the middle of the result set to see which is median
-  const midCount = valueCount / 2;
-  let median = 0;
-  let currentDistance = 0;
-  for (let i = 0; i < data.length; i += 1) {
-    currentDistance += data[i];
-    // We passed the mid point in this segment, we are sitting on the value.
-    if (currentDistance > midCount) {
-      median = i;
-      break;
-    }
-    // The median falls on the line between this segment and the next we have a bit of work to do.
-    // If this is a odd-length set (rare in this design) average this and the next.
-    // If there is gap between this value and the next, we need to measure it and pick the
-    // midpoint of the gap.
-    if (currentDistance === midCount) {
-      const lookAhead = data[i + 1];
-      if (lookAhead) {
-        // No gap, to the median is between here and the next.
-        median = (i + (i + 1)) / 2;
-        return median;
-      }
-      // We have a gap, so find next value.
-      let gap = 1;
-      let offset = 0;
-      for (let j = i + 2; j < data.length; j += 1) {
-        if (data[j] === 0) {
-          gap += 1;
-        } else {
-          // Found end of gap. The median is the middle of the gap.
-          offset = gap / 2;
-          break;
-        }
-      }
-      median = i + offset + 0.5;
-      return median;
-    }
-  }
-
-  return median;
-}
-
-/**
- * Calculates the standard deviation of the result list in compressed form.
- * The assumption here is that this is histogram-style data so each cell
- * is a count of results at that value.
- * Hat Tip: http://statisticshelper.com/standard-deviation-calculator-with-step-by-step-solution
- * @param {Array<number>} numberArray Histogram-style frequency array.
- * @returns {number} Sample standard deviation.
- */
-function getStandardDeviation(numberArray) {
-  // Calculate the average.
-  const sum = numberArray.reduce(
-    (total, currentValue, currentIndex) => total + (currentValue * currentIndex),
-  );
-  const valueCount = getValueCount(numberArray);
-  const avg = sum / valueCount;
-
-  // Calculate the standard deviation itself.
-  let sdPrep = 0;
-  for (let i = 0; i < numberArray.length; i += 1) {
-    sdPrep += ((i - avg) ** 2) * numberArray[i];
-  }
-
-  // Divide by n - 1
-  // This answer is the variance of the sample.
-  // Handle edge case where valueCount is 1 to avoid division by zero
-  const variance = valueCount > 1 ? sdPrep / (valueCount - 1) : 0;
-
-  const standardDev = Math.sqrt(variance);
-  return standardDev;
-}
-
-/**
- * Calculate the longest time the simulator may come up with.
- * @param {Array<Object>} tasks The list of tasks being tested.
- * @param {boolean} useCost Calculates the cost instead of time boundary.
- * @returns {number} Combined worst-case upper bound across all tasks.
- */
-function calculateUpperBound(tasks, useCost = false) {
-  let total = 0;
-  let multiplier = 1;
-  tasks.forEach((row) => {
-    if (useCost) {
-      multiplier = row.Cost;
-    }
-    total += (taskUpperBound(row.Max, row.Confidence) * multiplier);
-  });
-  return total;
-}
-
-/**
- * Builds the histogram graph, and returns an svg from D3. When there is lots of
- * data it will automatically convert to use a xy scatter graph instead of a bar graph.
- * @param {HTMLElement} targetNode The DOM element to insert the graph into.
- * @param {Array} list List of values to display
- * @param {number} min Smallest value
- * @param {number} max Largest value
- * @param {number} median Median
- * @param {number} stdDev Standard Deviation of values.
- * @param {boolean} limitGraph Limits the display to two standard deviations.
- */
-function buildHistogram(targetNode, list, min, max, median, stdDev, xLabel, limitGraph) {
-  // Remove and existing graphs
-  targetNode.innerHTML = '';
-
-  // The number of points before it switches to using a line graph.
-  const {
-    barCutoff, width: imageWidth, height: imageHeight, margin,
-  } = GRAPH_CONFIG.histogram;
-
-  // Image Margins
-  const binMargin = 0.2;
-
-  // Set outer bounds of graph.
-  let minBin = min;
-  let maxBin = max;
-
-  // Trim the array to just hold cells in the range of results.
-  // If limit graph is set, just show two standard deviations on the graph.
-  if (limitGraph) {
-    maxBin = median + (stdDev * 2) < max ? median + (stdDev * 2) : max;
-    minBin = median - (stdDev * 2) > min ? median - (stdDev * 2) : min;
-  }
-  const data = list.filter((e, i) => (i >= minBin && i <= maxBin));
-
-  // We need to round the median and standard deviation and find the
-  // index we expect for the bars.
-  const medianIndex = Math.round(median - minBin);
-  const stdDevOffset = Math.round(stdDev);
-  const stdDevLowIndex = medianIndex - stdDevOffset;
-  const stdDevHighIndex = medianIndex + stdDevOffset;
-
-  // whitespace on either side of the bars
-  const width = imageWidth - margin.left - margin.right;
-  const height = imageHeight - margin.top - margin.bottom;
-
-  // Set the limits of the x axis
-  const xMin = minBin - 1;
-  const xMax = maxBin + 1;
-
-  // Determine if we should use a bar graph or scatter plot.
-  const useBars = (xMax - xMin) < barCutoff;
-
-  // Set the max range of the y axis.
-  // The data array can be quite large so standard Math.Max approaches can fail.
-  const sortedData = [...data].sort((a, b) => a - b);
-  const yMax = sortedData[sortedData.length - 1];
-
-  // This scale is for determining the widths of the histogram bars
-  const x = d3.scaleLinear()
-    .domain([0, (xMax - xMin)])
-    .range([0, width]);
-
-  // Scale for the placement of the bars
-  const x2 = d3.scaleLinear()
-    .domain([xMin, xMax])
-    .range([0, width]);
-
-  const y = d3.scaleLinear()
-    .domain([0, yMax])
-    .range([height, 0]);
-
-  const xAxis = d3.axisBottom().scale(x2);
-  const yAxis = d3.axisLeft().scale(y).ticks(8);
-
-  // Put the graph in the histogram div.
-  const svg = d3.select(targetNode).append('svg')
-    .attr('role', 'img')
-    .attr('aria-label', `Histogram showing distribution of ${xLabel}. Median: ${median.toFixed(2)}, Standard Deviation: ${stdDev.toFixed(2)}, Range: ${min} to ${max}`)
-    .attr('width', width + margin.left + margin.right)
-    .attr('height', height + margin.top + margin.bottom)
-    .style('opacity', 0)
-    .append('g')
-    .attr('transform', `translate(${margin.left},${margin.top})`);
-
-  const renderTransition = d3.transition().duration(250).ease(d3.easeCubicOut);
-
-  // Add the x axis and x-label.
-  svg.append('g')
-    .attr('class', 'x axis')
-    .attr('transform', `translate(0,${height})`)
-    .call(xAxis);
-  svg.append('text')
-    .attr('class', 'xLabel')
-    .attr('text-anchor', 'middle')
-    .attr('x', width / 2)
-    .attr('y', height + margin.bottom - 10)
-    .text(xLabel);
-
-  // Add the y axis and y-label.
-  svg.append('g')
-    .attr('class', 'y axis')
-    .attr('transform', 'translate(0,0)')
-    .call(yAxis);
-  svg.append('text')
-    .attr('class', 'yLabel')
-    .attr('y', 0 - margin.left) // x and y switched due to rotation.
-    .attr('x', 0 - (height / 2))
-    .attr('dy', '1em')
-    .attr('transform', 'rotate(-90)')
-    .style('text-anchor', 'middle')
-    .text('Frequency');
-
-  // Set up the bars.
-  if (useBars) {
-    const bar = svg.selectAll('.bar')
-      .data(data)
-      .enter().append('g')
-      .attr('class', (d, i) => {
-        if (i === medianIndex) {
-          return 'bar median';
-        } if (i > stdDevLowIndex && i < stdDevHighIndex) {
-          return 'bar stdDev';
-        }
-        return 'bar';
-      })
-      .attr('transform', (d, i) => `translate(${x2(i + minBin)},${height})`);
-
-    bar.transition(renderTransition)
-      .attr('transform', (d, i) => `translate(${x2(i + minBin)},${y(d)})`);
-
-    // Add rectangles of correct size at correct location.
-    bar.append('rect')
-      .attr('x', x(binMargin))
-      .attr('width', x(2 * binMargin))
-      .attr('height', 0)
-      .transition(renderTransition)
-      .attr('height', (d) => height - y(d));
-  } else {
-    // Use Scatter plot instead of bars
-    const points = svg.selectAll('dot')
-      .data(data)
-      .join('circle')
-      .attr('cx', (d, i) => x2(i + minBin))
-      .attr('cy', (d) => y(d))
-      .attr('opacity', 0)
-      .attr('r', (d, i) => {
-        if (i === medianIndex) {
-          return 3;
-        }
-        return 1;
-      })
-      .attr('class', (d, i) => {
-        if (i === medianIndex) {
-          return 'graphXY median';
-        } if (i > stdDevLowIndex && i < stdDevHighIndex) {
-          return 'graphXY stdDev';
-        }
-        return 'graphXY';
-      });
-
-    points.transition(renderTransition)
-      .attr('opacity', 1);
-
-    // Draw kernel density estimate curve
-    const kdeData = calculateKDE(data, minBin, maxBin);
-    const kdeStep = (maxBin - minBin) / kdeData.length;
-
-    const lineGenerator = d3.line()
-      .x((d, i) => x2(minBin + (i * kdeStep)))
-      .y((d) => y(d))
-      .curve(d3.curveCardinal.tension(0.5));
-
-    svg.append('path')
-      .datum(kdeData)
-      .attr('class', 'kde-curve')
-      .attr('fill', 'none')
-      .attr('stroke', '#ef4444')
-      .attr('stroke-width', 2.5)
-      .attr('opacity', 0)
-      .attr('d', lineGenerator)
-      .transition(renderTransition)
-      .attr('opacity', 0.8);
-  }
-
-  d3.select(targetNode)
-    .select('svg')
-    .transition(renderTransition)
-    .style('opacity', 1);
-}
-
-/**
- * Builds a lightweight histogram preview for progressive simulation updates.
- * Uses fixed bucket counts and no KDE for fast redraw performance.
- * @param {HTMLElement} targetNode The DOM element to insert the graph into.
- * @param {Array<number>} list List of values to display.
- * @param {number} min Smallest value.
- * @param {number} max Largest value.
- * @param {string} xLabel X axis label.
- */
-function buildHistogramPreview(targetNode, list, min, max, xLabel) {
-  if (min < 0 || max < min) {
-    return;
-  }
-
-  const {
-    width: imageWidth, height: imageHeight, margin,
-  } = GRAPH_CONFIG.histogram;
-  const width = imageWidth - margin.left - margin.right;
-  const height = imageHeight - margin.top - margin.bottom;
-
-  const valueRange = (max - min) + 1;
-  const rawMaxBuckets = GRAPH_CONFIG.histogram.maxBuckets;
-  const maxBuckets = (Number.isFinite(rawMaxBuckets) && rawMaxBuckets > 0)
-    ? Math.floor(rawMaxBuckets)
-    : 120; // fallback if config value is invalid
-  const bucketCount = Math.max(1, Math.min(maxBuckets, valueRange));
-  const bucketSize = Math.max(1, Math.ceil(valueRange / bucketCount));
-  const buckets = new Array(bucketCount).fill(0);
-
-  for (let i = min; i <= max; i += 1) {
-    const bucketIndex = Math.min(Math.floor((i - min) / bucketSize), bucketCount - 1);
-    buckets[bucketIndex] += list[i] || 0;
-  }
-
-  let yMax = 0;
-  for (const bucketValue of buckets) {
-    if (bucketValue > yMax) {
-      yMax = bucketValue;
-    }
-  }
-
-  if (yMax < 1) {
-    return;
-  }
-
-  const x = d3.scaleLinear()
-    .domain([0, bucketCount])
-    .range([0, width]);
-
-  const x2 = d3.scaleLinear()
-    .domain([min, max])
-    .range([0, width]);
-
-  const y = d3.scaleLinear()
-    .domain([0, yMax])
-    .range([height, 0]);
-
-  const xAxis = d3.axisBottom().scale(x2).ticks(8);
-  const yAxis = d3.axisLeft().scale(y).ticks(8);
-
-  const container = d3.select(targetNode);
-  let svg = container.select('svg.preview-svg');
-
-  if (svg.empty()) {
-    targetNode.innerHTML = '';
-    svg = container.append('svg')
-      .attr('class', 'preview-svg')
-      .attr('role', 'img');
-
-    const root = svg.append('g')
-      .attr('class', 'preview-root');
-
-    root.append('g').attr('class', 'x axis');
-    root.append('text')
-      .attr('class', 'xLabel')
-      .attr('text-anchor', 'middle');
-
-    root.append('g').attr('class', 'y axis');
-    root.append('text')
-      .attr('class', 'yLabel')
-      .attr('dy', '1em')
-      .attr('transform', 'rotate(-90)')
-      .style('text-anchor', 'middle')
-      .text('Frequency');
-  }
-
-  svg
-    .attr('aria-label', `Histogram preview showing distribution of ${xLabel}. Range: ${min} to ${max}`)
-    .attr('width', width + margin.left + margin.right)
-    .attr('height', height + margin.top + margin.bottom);
-
-  const root = svg.select('g.preview-root')
-    .attr('transform', `translate(${margin.left},${margin.top})`);
-
-  const barWidth = Math.max((width / bucketCount) - 1, 1);
-  const renderTransition = d3.transition().duration(180).ease(d3.easeCubicOut);
-
-  root.select('.x.axis')
-    .attr('transform', `translate(0,${height})`)
-    .transition(renderTransition)
-    .call(xAxis);
-
-  root.select('.xLabel')
-    .attr('x', width / 2)
-    .attr('y', height + margin.bottom - 10)
-    .text(xLabel);
-
-  root.select('.y.axis')
-    .transition(renderTransition)
-    .call(yAxis);
-
-  root.select('.yLabel')
-    .attr('y', 0 - margin.left)
-    .attr('x', 0 - (height / 2));
-
-  const bars = root.selectAll('rect.preview-bar')
-    .data(buckets, (d, i) => i);
-
-  bars.join(
-    (enter) => enter.append('rect')
-      .attr('class', 'preview-bar')
-      .attr('x', (d, i) => x(i) + 0.5)
-      .attr('width', barWidth)
-      .attr('y', height)
-      .attr('height', 0)
-      .call((selection) => selection.transition(renderTransition)
-        .attr('y', (d) => y(d))
-        .attr('height', (d) => height - y(d))),
-    (update) => update
-      .call((selection) => selection.transition(renderTransition)
-        .attr('x', (d, i) => x(i) + 0.5)
-        .attr('width', barWidth)
-        .attr('y', (d) => y(d))
-        .attr('height', (d) => height - y(d))),
-    (exit) => exit
-      .call((selection) => selection.transition(renderTransition)
-        .attr('y', height)
-        .attr('height', 0)
-        .remove()),
-  );
-}
-
-/**
- * Builds a compact histogram visualization for a single task row.
- * @param {HTMLElement} targetNode Graph container for one task row.
- * @param {Array<number>} list Task histogram data.
- * @param {number} min Minimum simulated value.
- * @param {number} max Maximum simulated value.
- * @param {string} taskName Task display name.
- */
-function buildTaskRowHistogram(targetNode, list, min, max, taskName) {
-  targetNode.innerHTML = '';
-
-  if (min < 0 || max < min) {
-    return;
-  }
-
-  const {
-    width: graphWidth, height: graphHeight, maxBuckets, gap,
-  } = GRAPH_CONFIG.miniGraph;
-  const valueRange = max - min + 1;
-  const bucketCount = Math.min(maxBuckets, valueRange);
-  const bucketSize = Math.ceil(valueRange / bucketCount);
-  const buckets = new Array(bucketCount).fill(0);
-
-  for (let i = min; i <= max; i += 1) {
-    const bucketIndex = Math.min(Math.floor((i - min) / bucketSize), bucketCount - 1);
-    buckets[bucketIndex] += list[i];
-  }
-
-  let peak = 0;
-  for (const value of buckets) {
-    if (value > peak) {
-      peak = value;
-    }
-  }
-
-  if (peak === 0) {
-    return;
-  }
-
-  const svgNs = 'http://www.w3.org/2000/svg';
-  const svg = document.createElementNS(svgNs, 'svg');
-  svg.setAttribute('width', String(graphWidth));
-  svg.setAttribute('height', String(graphHeight));
-  svg.setAttribute('viewBox', `0 0 ${graphWidth} ${graphHeight}`);
-  svg.setAttribute('role', 'img');
-  svg.setAttribute('aria-label', `Task outcome histogram for ${taskName || 'task'}`);
-
-  const barWidth = Math.max((graphWidth / bucketCount) - gap, 1);
-
-  for (let i = 0; i < buckets.length; i += 1) {
-    const bucketValue = buckets[i];
-    const barHeight = Math.max((bucketValue / peak) * graphHeight, 1);
-    const rect = document.createElementNS(svgNs, 'rect');
-    rect.setAttribute('x', String(i * (barWidth + gap)));
-    rect.setAttribute('y', String(graphHeight - barHeight));
-    rect.setAttribute('width', String(barWidth));
-    rect.setAttribute('height', String(barHeight));
-    rect.setAttribute('class', 'task-row-mini-bar');
-    svg.appendChild(rect);
-  }
-
-  targetNode.appendChild(svg);
-}
+import {
+  getRandom,
+  taskUpperBound,
+  taskLowerBound,
+  generateEstimate,
+  getValueCount,
+  calculateKDE,
+  getMedian,
+  getStandardDeviation,
+  calculateUpperBound,
+} from './stats';
+import {
+  GRAPH_CONFIG,
+  GRAPH_CONFIG_DEFAULTS,
+  buildHistogram,
+  buildHistogramPreview,
+  buildTaskRowHistogram,
+} from './charts';
 
 /**
  * Yields execution to allow UI updates between simulation batches.
@@ -888,8 +193,6 @@ function runSimulationCore(passes, data, callbacks = {}, hoursPerTimeUnit = 1) {
       const processedPasses = endIndex;
       const hasMoreBatches = endIndex < totalPasses;
 
-      // Call progress callback
-      // Calculate current statistics for progressive display
       const currentTimeMedian = getMedian(times);
       const currentTimeSd = getStandardDeviation(times);
       const currentCostMedian = getMedian(costs);
@@ -1010,15 +313,11 @@ function fibonacciToVelocityDays(fibonacci, pointsPerSprint, sprintLengthDays) {
   const points = parseFloat(pointsPerSprint);
   const days = parseFloat(sprintLengthDays);
 
-  // Validate inputs
   if (fib <= 0 || points <= 0 || days <= 0) {
     return { min: 0, max: 0 };
   }
 
-  // Calculate points per day ratio
   const pointsPerDay = points / days;
-
-  // Base calculation: story points / velocity
   const baseDays = fib / pointsPerDay;
 
   // Apply variance: ±30% to account for uncertainty, rounded to whole days
@@ -1028,12 +327,16 @@ function fibonacciToVelocityDays(fibonacci, pointsPerSprint, sprintLengthDays) {
   return { min, max };
 }
 
+// Re-export everything from sub-modules so existing callers (index.js, tests) need no changes.
 export {
-  runSimulation,
-  runSimulationProgressive,
-  buildHistogramPreview,
+  GRAPH_CONFIG,
+  GRAPH_CONFIG_DEFAULTS,
   buildHistogram,
+  buildHistogramPreview,
   buildTaskRowHistogram,
+} from './charts';
+
+export {
   getRandom,
   getValueCount,
   getMedian,
@@ -1041,8 +344,11 @@ export {
   calculateKDE,
   taskUpperBound,
   taskLowerBound,
+} from './stats';
+
+export {
+  runSimulation,
+  runSimulationProgressive,
   fibonacciToCalendarDays,
   fibonacciToVelocityDays,
-  GRAPH_CONFIG,
-  GRAPH_CONFIG_DEFAULTS,
 };
